Move genshi._six.moves.builtins and genshi._six.exec_ to genshi.compat.
